Improvements to AnimPlayer and some tidy up - Fixed snapping that could occur with animations were first and last frames need to blend - Added ability to blend to an arbitrary frame

+++ Engine/Engine.h
@@ -19,7 +19,7 @@ namespace TombForge
     struct DebugData
     {
         double lastFpsUpdate{};
-        float fps{};
+        int fps{};
         int framesThisSecond{};
     };
